Fix the calculation of a schema size when partitioned tables present
In the presence of partitioned tables, the size of their files was taken into
account twice when calculating the schema size.

Add the own_file column to table_files to distinguish its own table files from
files of its partitions. Use this new column to add the size of each file only
once when schema size is calculated.

In addition, replace UNION with UNION ALL to simplify the query plan, as
duplicate rows are not possible.